Move SharesDatabase into its own file
